+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Provide an option for downstream SDKs to attach a view hierarchy file. ([#1191](https://github.com/getsentry/sentry-native/pull/1191))
 
+**Fixes**:
+
+- Provide a more defensive automatic thread stack guarantee. ([#1196](https://github.com/getsentry/sentry-native/pull/1196))
+
 ## 0.8.3
 
 **Features**: